primitives: Move script module tests to separate file
Currently we have test modules in three files within the `script`
module. Its not immediately what is tested where. We can put all the
tests in one file making it easier to see what is going no. Also
because `ScriptBuf` derefs to `Script` testing one tests the other.
